Other Info navigation bar to animate downwards
