+++ src/components/AnalysisPage.js
@@ -153,7 +153,8 @@ AnalysisPage.getInitialProps = async ({
         );
 
         const topics = await Promise.all(
-          sectionTopics.map(async topic => {
+          sectionTopics.map(async t => {
+            const topic = t.profile_section_topic || t;
             if (topic.post_content === '') {
               topic.type = 'carousel_topic'; // eslint-disable-line no-param-reassign
               // add another backend call to fetch the carousel_topic
